Few more enhancements done:
1. pdf.worker.js is bundled separately. This is required as a separate file
by the pdf.js

2. Corrected the logic for comment thumbnail generation.

3. Sometimes the PDF pages were being rendered in incorrect order. This
happens as the pdf.getPage is async in nature. When calling it in a for
loop for all pages, the order in which the callback is called is not
predictable. Changed the logic to first generate all the parent div
elements that will hold the per page canvas elements. As a second step,
each canvas element is created and appended to its corresponding parent
div.